// web/src/app/modules/roosid/model/action.ts
export enum Action {
  GET_CURRENT = 'GET_CURRENT',
  SET_ACTIVE_ROUND = 'SET_ACTIVE_ROUND',
  SET_INACTIVE_ROUND = 'SET_INACTIVE_ROUND',
  TOGGLE_ANSWER = 'TOGGLE_ANSWER',
  ADD_CURRENT_SCORE_TO_TEAM = 'ADD_CURRENT_SCORE_TO_TEAM',
  RESET_SCORE = 'RESET_SCORE',
  ADD_CROSS = 'ADD_CROSS',
  RESET_CROSS = 'RESET_CROSS',
  PLAY_SHORT_THEME = 'PLAY_SHORT_THEME',
  STOP_SHORT_THEME = 'STOP_SHORT_THEME',
  PLAY_INTRO = 'PLAY_INTRO',
  STOP_INTRO = 'STOP_INTRO',
  SET_SCORE_TO_WIN = 'SET_SCORE_TO_WIN',
  RESET_CROSS_ALL = 'RESET_CROSS_ALL',
  SET_ROUND_WINNER = 'SET_ROUND_WINNER',
  RESTART_GAME = 'RESTART_GAME',
  CHANGE_BUFFER = 'CHANGE_BUFFER',
  CHANGE_ZOOM = 'CHANGE_ZOOM',
  SET_ANSWERING_FOR_ROUND = 'SET_ANSWERING_FOR_ROUND',
  SET_FAST_MONEY_ACTIVE = 'SET_FAST_MONEY_ACTIVE',
  SET_FAST_MONEY_INACTIVE = 'SET_FAST_MONEY_INACTIVE',
  SAVE_PLAYER_ANSWERS = 'SAVE_PLAYER_ANSWERS',
  TOGGLE_FAST_MONEY_ANSWER = 'TOGGLE_FAST_MONEY_ANSWER'
}